requirements.txt
alive-progress
termcolor
python-dotenv